chore: ganti float jadi int

+++ app/src/main/java/nimonscooked/entity/Entity.java
@@ -1,5 +1,7 @@
 package nimonscooked.entity;
 
+import java.awt.image.BufferedImage;
+
 public abstract class Entity{
     protected String id;
     protected String name;
@@ -16,7 +18,7 @@ public abstract class Entity{
         this.y = y;
     }
 
-    public void moveTo(float X, float Y) {
+    public void moveTo(int X, int Y) {
         this.x += X;
         this.y += Y;
     }
@@ -37,19 +39,19 @@ public abstract class Entity{
         this.name = name;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
