+++ app/src/main/java/nimonscooked/entity/item/ingredient/Meat.java
@@ -3,7 +3,7 @@ package nimonscooked.entity.item.ingredient;
 import nimonscooked.enums.IngredientState;
 
 public class Meat extends Ingredient{
-    public Meat(String id, float x, float y, IngredientState currentState) {
+    public Meat(String id, int x, int y, IngredientState currentState) {
         super(id, "Meat", x, y, currentState);
     }
